Improve parsing and error messages
* Add test for dungeon-parsing code
* Use our own custom optionalListField function instead of oneOf,
  which catches inner failures.

`oneOf [field "x" a, b]` falls back to b if `field "x" a`
generates _any_ errors, (e.g. because we expected a to be
a list and it was a string).

But if `a` is invalid, I still want an error.
I only want to use the default if there's no field `"x"` at all.

Hence this new `optionalListField` Decoder.

And for some reason, adding `optionalListField` broke
recursive decoding of regions, so that's re-written as well.