Make detailed exit format easier to parse
Previously exits could be described like `secret: 42`
Now, we expect a format like:

    type: secret
    to: 42
    ... possibly other door details ...

This makes extra information in the door definition unambiguous
(e.g. `signs: 'one brick is polished smooth'` for a secret door).

Previously we could parse this as a connection of type "signs"
to the room keyed as "one brick is ...", since the key for the
door type could be any text.

We could have instead fixed this by requiring the door definition
itself to appear first, but that probably wouldn't be reliable
unless we also required an ordered hash (which would add extra
boilerplate to no purpose)